renamed and updated references to src/link.ts
So package now builds on systems with case sensitive file systems
