+++ lib/link.d.ts
